==> 03_files_mapping.Rmd <==
---
title: "03_files_mapping.Rmd"
author: "Thomas Naake"
date: "05/22/2021"
fig_width: 15
fig_height: 10
output:
    html_document:
    toc: true
toc_depth: 3
toc_float: TRUE
theme: united
number_sections: true
highlight: tango
---


```{r}
## read the files that contain the mapping information
setwd("~/Projects/20210522_molNetworking_swM")

options(stringsAsFactors = FALSE)
cid_pos <- read.table("Pos_cid_Peak_SNP_locus_LD_candidate.genes.csv", 
    header = TRUE, sep = ",", quote = "`")
cid_neg <- read.table("Neg_cid_Peak_SNP_locus_LD_candidate.genes.csv", 
    header = TRUE, sep = ",", quote = "`")
hcd_pos <- read.table("Pos_hcd_Peak_SNP_locus_LD_candidate.genes.csv", 
    header = TRUE, sep = ",", quote = "`")
hcd_neg <- read.table("Neg_hcd_Peak_SNP_locus_LD_candidate.genes.csv", 
    header = TRUE, sep = ",", quote = "`")

setwd("~/Projects/20210522_molNetworking_swM/results_MS1/")
tmp <- read.table("peaklist_pos_swM_hcd_map.txt", sep = "\t")
```

```{r}
################################################################################
## function to write the mapping results (genes to a list)
write_to_list <- function(mapping, type = "HCD") {
    uniq_trait <- unique(mapping[, "Trait"])
    gene_l <- vector("list", length(uniq_trait))
    ##uniq_trait_cut <- unlist(lapply(strsplit(uniq_trait, split = type), "[", 2))
    names(gene_l) <- paste(type, uniq_trait, sep = "_")
    for (i in 1:length(gene_l)) {
        inds <- which(mapping[, "Trait"] == uniq_trait[i])
        gene_l[[i]] <- mapping[inds, "Gene.stable.ID"]
    }
    gene_l
}
```

```{r}
## apply the function: create a list containing the genes mapped to a metabolite
cid_pos_l <- write_to_list(cid_pos, "swM_CID")
names(cid_pos_l) <- gsub(names(cid_pos_l), pattern = "PCID", replacement = "")
cid_neg_l <- write_to_list(cid_neg, "swM_CID")
names(cid_neg_l) <- gsub(names(cid_neg_l), pattern = "NCID", replacement = "")
hcd_pos_l <- write_to_list(hcd_pos, "swM_HCD")
names(hcd_pos_l) <- gsub(names(hcd_pos_l), pattern = "PHCD", replacement = "")
hcd_neg_l <- write_to_list(hcd_neg, "swM_HCD")
names(hcd_neg_l) <- gsub(names(hcd_neg_l), pattern = "NHCD", replacement = "")

## load the spectra
load("../results_MS2/ms2_spectra_MK_cid_pos.RData")
load("../results_MS2/ms2_spectra_MK_hcd_pos.RData")
load("../results_MS2/ms2_spectra_ML_cid_pos.RData")
load("../results_MS2/ms2_spectra_ML_hcd_pos.RData")
load("../results_MS2/ms2_spectra_swM_cid_pos.RData")
load("../results_MS2/ms2_spectra_swM_hcd_pos.RData")
load("../results_MS2/ms2_similarityMat_swM_hcd_pos.RData")
```

Create two matrices: 

- in the first store if there is an overlap between the mapped genes
- in the second store the similarity based on NDP if there is an overlap 
   between the mapped genes


Information on the mapping between the different data type:
 - `hcd_pos`: mapping result of the traits (metabolites)
 - `peaklist_pos_swM_hcd_map` (`tmp`): MS1 with information on mz, rt, intensities 
      and corresponding MS2 spectra
 - `similarityMat_swM_hcd_pos`: similarity matrix of MS2 features, rownames and 
    colnames correspond to the spectra names

```{r eval=FALSE, echo=FALSE}
similarityMat_swM_hcd_pos[1:10,1:10]
hcd_pos[1:5, 1:5] ## PCHCD1828 --> refers to swM_HCD_1828 in peaklist_...[, "spectra"] (tmp)
tmp[tmp[, "spectra"] == "swM_HCD_1828", ] ## --> this is spectra 1203 in similarityMat
length(unique(tmp[, "spectra"]))
similarityMat_swM_hcd_pos["swM_HCD_1828", 1:10]
```

```{r}
## create function to find overlapping loci and similarities between the
## mapped MS1 peaks
create_overlap_matrices <- function(l = hcd_pos_l, 
        similarityMat = similarityMat_MK_hcd_pos) {
    mat_sim <- mat <- matrix(nrow = length(l), ncol = length(l), 
    dimnames = list(names(l), names(l)))
    mat[is.na(mat)] <- ""
    mat_sim[is.na(mat_sim)] <- 0
    
    ## iterate through rows and cols of mat (these refer to the mapped MS1 
    ## spectra with loci) and check if there is an overlap in the mapped genes
    ## (MS1/MS2 peaklist)
    for (i in 1:nrow(mat)) {
        row_i <- rownames(mat)[i] ## refers to XYZ in peaklist_MS1[, "spectra"]
    
        for (j in 1:ncol(mat)) {
            col_j <- colnames(mat)[j]
        
            ## obtain the similarity between the two MS2 features i and j, 
            ## which were mapped to a genetic locus/loci
            ## similarityMat_... is symmetric, thus it doesnt matter if to take 
            ## [i, j] or [j, i]
            if (all(c(row_i, col_j) %in% rownames(similarityMat))) {
                sim_ij <- similarityMat[row_i, col_j]
            } else {
                sim_ij <- NA
            }
        
            ## get the genes that are mapped to MS2 features i and j
            genes_mapped_i <- l[[row_i]]
            genes_mapped_j <- l[[col_j]]
        
            ## check if there is an overlap between genes_mapped_i and 
            ## genes_mapped_j; if so, write the overlap in matrix mat and the 
            ## similarity in the matrix mat_sim
            if (any(genes_mapped_i %in% genes_mapped_j)) {
                mat_sim[i, j] <- mat_sim[j, i] <- sim_ij
                mat[i, j] <- mat[j, i] <- 
                    paste(intersect(genes_mapped_i, genes_mapped_j), 
                        collapse = ",")
            }
        }
    }
    list(loci_matrix = mat, similarity_matrix = mat_sim)
}

hcd_pos_mat_l <- create_overlap_matrices(l = hcd_pos_l, similarityMat = similarityMat_swM_hcd_pos)
mat <- hcd_pos_mat_l[["loci_matrix"]]
mat_sim <- hcd_pos_mat_l[["similarity_matrix"]]

```


# Visualize with igraph
```{r}
library(igraph)

mat_sim_red <- mat_sim
mat_sim_red[mat_sim_red < 0.05] <- 0
g <- graph_from_adjacency_matrix(mat_sim_red, mode = "undirected", weighted = TRUE, diag = FALSE)
plot(g, vertex.size = 0.1, vertex.label = NA)

## write the files 
write.table(mat_sim, file = "swM_hcd_pos_similarity_GWAS_adjacency.txt", 
    quote = FALSE, row.names = TRUE, col.names = TRUE)
write.table(mat, file = "swM_hcd_pos_overlap_GWAS_adjacency.txt", 
    quote = FALSE, row.names = TRUE, col.names = TRUE)
```

## check some of the metabolites

quinic acid
```{r}
mat_sim["swM_HCD_88", c("swM_HCD_500", "swM_HCD_580", "swM_HCD_1298")]
mat["swM_HCD_88", c("swM_HCD_500", "swM_HCD_580", "swM_HCD_1298")]
```

Kaempferol-3-O-rutinoside
```{r}
mat_sim["swM_HCD_1667", 
        c("swM_HCD_1668", "swM_HCD_1613", "swM_HCD_1555", "swM_HCD_1823")]
mat["swM_HCD_1667", 
        c("swM_HCD_1668", "swM_HCD_1613", "swM_HCD_1555", "swM_HCD_1823")]
```

Naringenin
```{r}
mat_sim["swM_HCD_1664", 
        c("swM_HCD_1667", "swM_HCD_1668", "swM_HCD_1613", "swM_HCD_1555", "swM_HCD_1823")]
mat["swM_HCD_1664", 
        c("swM_HCD_1667", "swM_HCD_1668", "swM_HCD_1613", "swM_HCD_1555", "swM_HCD_1823")]

```
